Remove moved source

+++ docs/_scripts/build_docs/fiddle.ts
@@ -16,10 +16,6 @@ import fs from "fs-extra";
 
 const __dirname = path.resolve("./docs/_scripts/build_docs");
 
-const FIDDLE_IN_PATH = path.join(
-  __dirname,
-  "../../../demo/malloy-duckdb-wasm/docs"
-);
 const FIDDLE_OUT_PATH = path.join(__dirname, "../../fiddle");
 const FIDDLE_INDEX_FILE = path.join(FIDDLE_OUT_PATH, "index.html");
 
@@ -29,11 +25,13 @@ layout: fiddle
 `;
 
 export async function copyFiddle(): Promise<void> {
-  await fs.rm(FIDDLE_OUT_PATH, { recursive: true, force: true });
-  await fs.copy(FIDDLE_IN_PATH, FIDDLE_OUT_PATH);
+  if (process.env.MALLOY_FIDDLE_PATH) {
+    await fs.rm(FIDDLE_OUT_PATH, { recursive: true, force: true });
+    await fs.copy(process.env.MALLOY_FIDDLE_PATH, FIDDLE_OUT_PATH);
 
-  const index = fs.readFileSync(FIDDLE_INDEX_FILE, "utf-8");
-  const indexTemplate =
-    HEADER + index.replace(/<\/head>/g, `{% include ga.html %}\n  </head>`);
-  fs.writeFileSync(FIDDLE_INDEX_FILE, indexTemplate);
+    const index = fs.readFileSync(FIDDLE_INDEX_FILE, "utf-8");
+    const indexTemplate =
+      HEADER + index.replace(/<\/head>/g, `{% include ga.html %}\n  </head>`);
+    fs.writeFileSync(FIDDLE_INDEX_FILE, indexTemplate);
+  }
 }
